requirements.txt
appdirs==1.4.3
asgiref==3.2.7
bson==0.5.8
dataclasses==0.6
distlib==0.3.0
Django==2.2.11
djangorestframework==3.11.0
djongo==1.3.1
filelock==3.0.12
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.2.4
virtualenv==20.0.14
drf-yasg==1.17.1
django-allauth==0.41.0
django-rest-passwordreset==1.1.0
coverage

